requirements.txt
PyYAML==6.0.2
PyGithub==2.6.0
Jinja2==3.1.6
